refactor: Adjust import paths for consistency and clarity across modules
