+++ src/database/database.service.ts
@@ -1,16 +1,28 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, OnModuleDestroy } from '@nestjs/common';
 import { Pool } from 'pg';
 
 @Injectable()
-export class DatabaseService {
-  private pool = new Pool({
-    connectionString: process.env.DATABASE_URL,
-  });
+export class DatabaseService implements OnModuleDestroy {
+  private pool: Pool;
 
-  async query(query: string, params: any[] = []) {
-    console.log('DATABASE_URL', process.env.DATABASE_URL);
+  constructor() {
+    this.pool = new Pool({
+      connectionString: process.env.DATABASE_URL,
+    });
+  }
 
+  async query(query: string, params?: any[]) {
+    // console.log('DATABASE_URL', process.env.DATABASE_URL);
     const { rows } = await this.pool.query(query, params);
     return rows;
+  }
+
+  async close() {
+    // console.log('Closing database connection...');
+    await this.pool.end(); // Ensure connection is properly closed
+  }
+
+  async onModuleDestroy() {
+    await this.close();
   }
 }
