+++ src/urls/urls.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { DatabaseService } from './database.service';
 
 import { UrlMapper } from './mappers/url.mapper';
 import { UrlsService } from './urls.service';
 import { UrlsController } from './urls.controller';
+import { DatabaseService } from '../database';
 
 @Module({
   providers: [UrlsService, DatabaseService, UrlMapper],
