+++ tsconfig.json
@@ -9,7 +9,6 @@
     "target": "ES2021",
     "sourceMap": true,
     "outDir": "./dist",
-    "baseUrl": "./",
     "incremental": true,
     "skipLibCheck": true,
     "strictNullChecks": false,
@@ -18,12 +17,8 @@
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
     "paths": {
-      "@app/lib": [
-        "libs/lib/src"
-      ],
-      "@app/lib/*": [
-        "libs/lib/src/*"
-      ]
+      "@app/lib": ["../../libs/lib/src"],
+      "@app/lib/*": ["../../libs/lib/src/*"]
     }
   }
 }
